Fixed 'warning 208' by Southclaw
Fixed 'warning 208: function with tag result used before definition, forcing reparse' for these functions:
-bool:db_set_row_index
-bool:db_free_result_hook
-bool:db_set_synchronous

@Southclaw
Not entirely sure why this happens, but I know one fix for it is to put a "forward" line for the affected functions. However, I'm not sure if that has any repercussions on performance or memory use.